requirements.txt
streamlit
streamlit-option-menu
SpeechRecognition
pydub
langdetect
gTTS
ffmpeg-python
google-generativeai
mysql-connector-python
tempfile
numpy
librosa
soundfile
matplotlib
noisereduce

